server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic_settings
python-dotenv
httpx
mysql-connector-python
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
alembic